agency payments

+++ database/migrations/2022_04_25_115358_create_agency_payments_table.php
@@ -18,7 +18,7 @@ class CreateAgencyPaymentsTable extends Migration
             $table->unsignedBigInteger('agency_id');
             $table->unsignedBigInteger('job_id');
             $table->string('customer_id');
-            $table->integer('amount');
+            $table->decimal('amount',5,2);
             $table->string('payment_status');
             $table->string('payment_mode')->nullable();
             $table->timestamps();
